requirements.txt
numpy==1.26.4
pandas==2.2.0
matplotlib==3.8.2
seaborn==0.13.2
scipy==1.12.0
scikit-learn==1.4.0
biopython==1.83
tqdm==4.66.1
plotly==5.20.0
openpyxl==3.1.2
requests==2.31.0
joblib==1.3.2
pathlib==1.0.1
PyYAML~=6.0.1